pyproject.toml
# ============================================================================
# Shokker Paint Booth — Python project metadata & tool configuration
# ----------------------------------------------------------------------------
# This file is intentionally tool-config-only; the bundled pyserver runtime
# ships prebuilt wheels under electron-app/server/pyserver/_internal/.
# Use this file for:
#   * `pip install -e .[dev]` in a dev checkout
#   * `black`, `ruff`, `pytest`, `mypy` configuration
#   * IDE awareness (Pylance, PyCharm) of project root
# ============================================================================

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# ----------------------------------------------------------------------------
# Project metadata (PEP 621)
# ----------------------------------------------------------------------------
[project]
name = "shokker-paint-booth"
description = "Shokker Paint Booth — automotive paint rendering engine with zone-based spec maps, pattern overlays, and real-time preview."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE.txt" }
authors = [
  { name = "Shokker Group", email = "ricky@shokkergroup.com" }
]
keywords = [
  "paint",
  "rendering",
  "pbr",
  "spec-map",
  "automotive",
  "iRacing",
  "electron",
  "flask",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: End Users/Desktop",
  "License :: Other/Proprietary License",
  "Operating System :: Microsoft :: Windows :: Windows 10",
  "Operating System :: Microsoft :: Windows :: Windows 11",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Scientific/Engineering :: Image Processing",
]
dynamic = ["version"]

# Runtime dependencies — the bundled pyserver already includes these, but this
# list lets `pip install -e .` work in a dev checkout.
dependencies = [
  "flask>=3.0",
  "flask-cors>=4.0",
  "numpy>=1.26",
  "scipy>=1.11",
  "Pillow>=10.0",
  "opencv-python-headless>=4.8",
  "requests>=2.31",
  "colorama>=0.4",
  "noise>=1.2",
]

# ----------------------------------------------------------------------------
# Optional dependencies — install with `pip install -e .[dev]`
# ----------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.4",
  "mypy>=1.8",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-xdist>=3.5",
  "pytest-timeout>=2.2",
  "ipython>=8.20",
  "rich>=13.7",
]
test = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-xdist>=3.5",
  "pytest-timeout>=2.2",
]
docs = [
  "mkdocs>=1.5",
  "mkdocs-material>=9.5",
]
gpu = [
  "torch>=2.1",
  "torchvision>=0.16",
]

[project.urls]
Homepage = "https://shokkergroup.com"
Repository = "https://github.com/shokkergroup/paint-booth"
Changelog = "https://github.com/shokkergroup/paint-booth/blob/main/CHANGELOG.md"

# ----------------------------------------------------------------------------
# Setuptools — treat the engine as the package root
# ----------------------------------------------------------------------------
[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { file = "VERSION.txt" }

[tool.setuptools.packages.find]
where = ["electron-app/server"]
include = ["engine*"]
exclude = ["tests*", "_archive*", "pyserver*"]

# ============================================================================
# Black — code formatter
# ============================================================================
[tool.black]
line-length = 120
target-version = ["py311", "py312", "py313"]
skip-string-normalization = false
skip-magic-trailing-comma = false
extend-exclude = '''
(
    /_archive/
  | /node_modules/
  | /dist/
  | /build/
  | /pyserver/_internal/
  | /__pycache__/
  | .*\.pyi$
)
'''

# ============================================================================
# Ruff — fast linter (replaces flake8, isort, pyupgrade)
# ============================================================================
[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [
  "_archive",
  "node_modules",
  "dist",
  "build",
  "electron-app/server/pyserver/_internal",
  "__pycache__",
  "*.pyi",
]

[tool.ruff.lint]
# Enable: pycodestyle (E/W), pyflakes (F), isort (I), pyupgrade (UP),
# flake8-bugbear (B), flake8-comprehensions (C4), flake8-simplify (SIM),
# pep8-naming (N), flake8-pytest-style (PT)
select = ["E", "W", "F", "I", "UP", "B", "C4", "SIM", "N", "PT"]
ignore = [
  "E501",   # line-too-long — black handles this
  "E741",   # ambiguous var name — we use `l`, `I` intentionally in math code
  "N803",   # argument name should be lowercase — shape tuples (H, W) are common
  "N806",   # variable in function should be lowercase — same reason
  "B008",   # function call in default — Flask routes use this pattern
  "SIM108", # ternary over if/else — sometimes hurts readability
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["N802", "N803"]  # test_What_When is fine
"__init__.py" = ["F401"]      # re-exports

[tool.ruff.lint.isort]
known-first-party = ["engine", "shokker"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ============================================================================
# Mypy — static type checker
# ============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
disallow_untyped_defs = false  # gradual typing — many engine fns still untyped
disallow_any_generics = false
exclude = [
  "_archive/",
  "node_modules/",
  "dist/",
  "build/",
  "electron-app/server/pyserver/_internal/",
  "tests/fixtures/",
]

[[tool.mypy.overrides]]
module = ["cv2.*", "noise.*", "scipy.*"]
ignore_missing_imports = true

# ============================================================================
# Pytest
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "-p",
  "no:cacheprovider",
  "--tb=short",
  "--showlocals",
  "-ra",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "gpu: marks tests requiring a CUDA GPU",
  "integration: marks tests that hit the live Flask server",
  "smoke: minimal-footprint sanity tests",
]
timeout = 300
filterwarnings = [
  "error",
  "ignore::DeprecationWarning:flask",
  "ignore::PendingDeprecationWarning",
  "ignore::ResourceWarning",
]

# ============================================================================
# Coverage
# ============================================================================
[tool.coverage.run]
branch = true
source = ["electron-app/server/engine"]
omit = [
  "*/tests/*",
  "*/_archive/*",
  "*/pyserver/_internal/*",
  "*/__init__.py",
]

[tool.coverage.report]
precision = 1
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
